requirements.txt
alembic==1.13.1
aniso8601==9.0.1
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Faker==25.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==1.11.1
Flask-Flash==2.0.0
flask-form==1.0.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
flask-marshmallow==0.10.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-SSLify==0.1.5
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
inflect==7.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
more-itertools==10.2.0
msgspec==0.18.6
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==1.26.4
protobuf==5.26.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.30
typeguard==4.2.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.1.2
